New imagebuilder, more logs, updated FAQ
Former-commit-id: cfa3398b3e377b5ed4b0c782c6f3d702dbdfb23f

+++ xiaomi.vacuum.gen1/firmwarebuilder/imagebuilder.sh
@@ -18,36 +18,64 @@
 # place english.pkg and v11_003077.pkg in this folder
 # place your authorized_keys in this folder
 #
+if [ ! -f /usr/bin/ccrypt ]; then
+    echo "Ccrypt not found! Please install it (e.g. by apt-get install ccrypt)"
+	exit 1
+fi
 if [ ! -f english.pkg ]; then
     echo "File english.pkg not found!"
+	exit 1
 fi
 
 if [ ! -f v11_003077.pkg ]; then
     echo "File v11_003077.pkg not found!"
+	exit 1
 fi
 
 if [ ! -f authorized_keys ]; then
     echo "File authorized_keys not found!"
+	exit 1
 fi
 
 if [[ $EUID -ne 0 ]]; then
 	echo "You must be a root user" 2>&1
 	exit 1
 else
+	# Generate SSH Host Keys
+	echo "Generate SSH Host Keys"
+	rm ssh_host_*
+	rm ssh_host_*.pub
+	ssh-keygen -N "" -t rsa -f ssh_host_rsa_key
+	ssh-keygen -N "" -t dsa -f ssh_host_dsa_key
+	ssh-keygen -N "" -t ecdsa -f ssh_host_ecdsa_key
+	ssh-keygen -N "" -t ed25519 -f ssh_host_ed25519_key
 	echo "decrypt soundfile"
 	ccrypt -d -K r0ckrobo#23456 english.pkg
 	mkdir sounds
 	cd sounds
 	echo "unpack soundfile"
-	tar -xzvf ../english.pkg
+	tar -xzf ../english.pkg
 	cd ..
 	echo "decrypt firmware"
 	ccrypt -d -K rockrobo v11_003077.pkg
 	echo "unpack firmware"
-	tar -xzvf v11_003077.pkg
+	tar -xzf v11_003077.pkg
+	if [ ! -f disk.img ]; then
+		echo "File disk.img not found! Decryption and unpacking was apparently unsuccessful."
+		exit 1
+	fi
 	mkdir image
 	mount -o loop disk.img image
 	cd image
+	echo "patch ssh host keys"
+	cat ../ssh_host_rsa_key > ./etc/ssh/ssh_host_rsa_key
+	cat ../ssh_host_rsa_key.pub > ./etc/ssh/ssh_host_rsa_key.pub
+	cat ../ssh_host_dsa_key > ./etc/ssh/ssh_host_dsa_key
+	cat ../ssh_host_dsa_key.pub > ./etc/ssh/ssh_host_dsa_key.pub
+	cat ../ssh_host_ecdsa_key > ./etc/ssh/ssh_host_ecdsa_key
+	cat ../ssh_host_ecdsa_key.pub > ./etc/ssh/ssh_host_ecdsa_key.pub
+	cat ../ssh_host_ed25519_key > ./etc/ssh/ssh_host_ed25519_key
+	cat ../ssh_host_ed25519_key.pub > ./etc/ssh/ssh_host_ed25519_key.pub
 	echo "disable SSH firewall rule"
 	sed -e '/    iptables -I INPUT -j DROP -p tcp --dport 22/s/^/#/g' -i ./opt/rockrobo/watchdog/rrwatchdoge.conf
 	echo "integrate SSH authorized_keys"
@@ -73,12 +101,18 @@ else
 	umount image
 	rm -rf image
 	rm -rf sounds
-	tar -czvf v11_003077_patched.pkg disk.img
+	echo "pack new firmware"
+	tar -czf v11_003077_patched.pkg disk.img
+	if [ ! -f v11_003077_patched.pkg ]; then
+		echo "File v11_003077_patched.pkg not found! Packing the firmware was unsuccessful."
+		exit 1
+	fi
 	rm -f disk.img
+	echo "encrypt firmware"
 	ccrypt -e -K rockrobo v11_003077_patched.pkg
 	mkdir -p output
 	mv v11_003077_patched.pkg.cpt output/v11_003077.pkg
 	cd output
-	md5sum v11_003077.pkg > v11_003077.md5
+	md5sum v11_003077.pkg > output/v11_003077.md5
 fi
